// todo/src/patterns.enum.ts
export enum Pattern {
  CREATE_TODO = 'TODO_CREATE_TODO',
  GET_TODO_WITH_ID = 'TODO_GET_TODO_WITH_ID',
  GET_TODOS_BY_USER = 'TODO_GET_TODOS_BY_USER',
  GET_TODOS_BY_TEAM = 'TODO_GET_TODOS_BY_TEAM',
  GET_TODOS_WITH_FILTER = 'TODO_GET_TODOS_WITH_FILTER',
  UPDATE_TODO = 'TODO_UPDATE_TODO',
  DELETE_TODO = 'TODO_DELETE_TODO',
  GET_USER_BY_ID = 'USER_GET_USER_BY_ID',
  GET_TEAM_BY_ID = 'TEAM_GET_TEAM_BY_ID',
  ADD_USER_TODO = 'USER_ADD_TODO',
  REMOVE_USER_TODO = 'USER_REMOVE_TODO',
  ADD_TEAM_TODO = 'TEAM_ADD_TODO',
  REMOVE_TEAM_TODO = 'TEAM_REMOVE_TODO',
}